requirements.txt
django
djangorestframework
numpy